voting

+++ src/apps/views/projects.go
@@ -6,6 +6,8 @@ import (
 	"sif/src/apps/auth"
 	"sif/src/apps/models"
 	"sif/src/apps/utils"
+	"sif/src/config"
+	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
@@ -33,9 +35,8 @@ func projectsGroup(router *gin.Engine) {
 	})
 
 	g.GET("/:id", func(c *gin.Context) {
-		id := c.Param("id")
 
-		p, err := models.GetProject(uuid.MustParse(id))
+		p, err := models.GetProject(uuid.MustParse(c.Param("id")), c.MustGet("user").(*models.User).ID)
 		if err != nil {
 			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 			return
@@ -47,6 +48,17 @@ func projectsGroup(router *gin.Engine) {
 		ctx, _ := c.Get("ctx")
 		identity, _ := c.Get("identity")
 
+		r, err := models.GetRoundLatestRound()
+		if err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+		now := time.Now()
+		if !config.Config.Debug && (now.Before(r.SubmissionStartAt) || now.After(r.SubmissionEndAt)) {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Submission period is closed"})
+			return
+		}
+
 		form := new(ProjectForm)
 		if err := c.ShouldBindJSON(form); err != nil {
 			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
@@ -54,7 +66,7 @@ func projectsGroup(router *gin.Engine) {
 		}
 		p := new(models.Project)
 		utils.Copy(form, p)
-		p.IdentityID = &identity.(*models.Identity).ID
+		p.IdentityID = identity.(*models.Identity).ID
 		if form.Status == nil {
 			p.Status = models.ProjectStatusActive
 		}
@@ -74,7 +86,27 @@ func projectsGroup(router *gin.Engine) {
 			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 			return
 		}
-		p := new(models.Project)
+		p, err := models.GetProject(uuid.MustParse(id), c.MustGet("user").(*models.User).ID)
+		if err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+		if p.IdentityID != c.MustGet("identity").(*models.Identity).ID {
+			c.JSON(http.StatusForbidden, gin.H{"error": "Forbidden"})
+			return
+		}
+
+		r, err := models.GetRound(p.RoundID)
+		if err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+		now := time.Now()
+		if !config.Config.Debug && (now.Before(r.SubmissionStartAt) || now.After(r.SubmissionEndAt)) {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Submission period is closed"})
+			return
+		}
+
 		utils.Copy(form, p)
 		p.ID = uuid.MustParse(id)
 		if err := p.Update(ctx.(context.Context)); err != nil {
@@ -84,15 +116,32 @@ func projectsGroup(router *gin.Engine) {
 		c.JSON(http.StatusOK, p)
 	})
 
-	g.DELETE("/:id", func(c *gin.Context) {
+	g.DELETE("/:id", OrganizationRequired(), func(c *gin.Context) {
 		ctx, _ := c.Get("ctx")
 		id := c.Param("id")
 
-		p, err := models.GetProject(uuid.MustParse(id))
+		p, err := models.GetProject(uuid.MustParse(id), c.MustGet("user").(*models.User).ID)
 		if err != nil {
 			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 			return
 		}
+
+		if p.IdentityID != c.MustGet("identity").(*models.Identity).ID {
+			c.JSON(http.StatusForbidden, gin.H{"error": "Forbidden"})
+			return
+		}
+
+		r, err := models.GetRound(p.RoundID)
+		if err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+		now := time.Now()
+		if !config.Config.Debug && (now.Before(r.SubmissionStartAt) || now.After(r.SubmissionEndAt)) {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Submission period is closed"})
+			return
+		}
+
 		if err := p.Delete(ctx.(context.Context)); err != nil {
 			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 			return
@@ -100,5 +149,40 @@ func projectsGroup(router *gin.Engine) {
 		c.JSON(http.StatusOK, gin.H{
 			"message": "success",
 		})
+	})
+
+	g.POST("/:id/votes", func(c *gin.Context) {
+		user := c.MustGet("user").(*models.User)
+		if user.IdentityVerifiedAt == nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "You must verify your identity before voting"})
+			return
+		}
+
+		project, err := models.GetProject(uuid.MustParse(c.Param("id")), user.ID)
+		if err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+
+		r, err := models.GetRound(project.RoundID)
+		if err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+		now := time.Now()
+		if !config.Config.Debug && (now.Before(r.VotingStartAt) || now.After(r.VotingEndAt)) {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "voting period is closed"})
+			return
+		}
+
+		vote := &models.Vote{
+			UserID:    user.ID,
+			ProjectID: project.ID,
+		}
+		if err := vote.Create(c.MustGet("ctx").(context.Context)); err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "already voted"})
+			return
+		}
+		c.JSON(http.StatusCreated, gin.H{"vote": vote})
 	})
 }
